Add description of SrcType to configuration file
The current documentation does not mention the `srcType` configuration parameter, but `ItemsProcessor.ProcessItem` checks for it. When using tile sets, where `src` is not set, this will lead to the erroneous error that `Src sample should be set`. 
I suggest the change to the documentation to clarify the need to set the `srcType` for tile sets.